requirements.txt
uvicorn[standard]==0.15.0
gunicorn==20.1.0
fastapi==0.79.0
pydantic
alembic==1.4.2
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
moviepy==1.0.3
numpy
deepface==0.0.75
opencv-contrib-python-headless
python-multipart
psycopg2-binary==2.9.5
sqlmodel==0.0.8
python-dotenv
tenacity
asyncpg
email-validator
python-jose
bcrypt
cryptography
passlib
pg8000
cloud-sql-python-connector
keras
tensorflow==2.11.0
